fix unit test errors

+++ app/src/main/java/com/asfoundation/wallet/ui/gamification/GamificationInteractor.kt
@@ -72,7 +72,7 @@ class GamificationInteractor @Inject constructor(
   ): ForecastBonusAndLevel {
     return ForecastBonusAndLevel(
       forecastBonus.status, fiatValue.amount, fiatValue.symbol
-      , level = forecastBonus.level.toInt()
+      , level = forecastBonus.level
     )
   }
